Move comments feature spec to their own file
Everything in this application currently evolves around guides and therefore
the guide spec has grown quite big. Extract comments-related specs to a new
spec file.
